+++ tests/conftest.py
@@ -9,5 +9,7 @@ def config():
     config.now = None
     config.which = "upcoming"
     config.display_mode = "list"
+    config.background_color = None
+    config.background_image = None
 
     return config